requirements.txt
# Общий requirements.txt для всех сервисов
# Собран из web/requirements.txt, bot/bot_prodazhi/requirements.txt и зависимостей из кода

# Web Framework
Django==5.2.8
asgiref==3.10.0

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
starlette==0.41.0
pydantic==2.7.4
pydantic_core==2.18.4

# Bot Framework
aiogram==3.6.0
aiogram==2.25.1  # Для совместимости со старым кодом (будет заменено)
aiohttp==3.9.5
aiofiles==23.2.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL
aiosqlite==0.19.0  # SQLite async (для миграции)
asyncpg==0.29.0  # PostgreSQL async

# HTTP Clients
httpx==0.27.0
requests==2.32.5
aiohappyeyeballs==2.6.1
aiosignal==1.4.0

# Utilities
python-dotenv==1.2.1
config==0.5.1
python-dateutil==2.9.0

# Scheduler
APScheduler==3.10.4

# Redis (для aiogram storage)
redis==5.0.1
hiredis==2.3.2

# Flask (для bot-api)
Flask==3.0.3
Werkzeug==3.0.3

# YooKassa
yookassa==3.0.0

# Google Sheets
gspread==6.2.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2

# Other dependencies
annotated-types==0.7.0
async-timeout==4.0.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.7.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
oauthlib==3.3.1
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests-oauthlib==2.0.0
rsa==4.9.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1

# Jinja2 (для FastAPI templates)
Jinja2==3.1.4
MarkupSafe==3.0.2
